Now with working country and header conditionlets.
