+++ src/view/components/rule-engine/rule-component.ts
@@ -133,8 +133,20 @@ class RuleComponent{
         fakeId: {
           id: 'fakeId',
           key: 'headerValue',
-          value: 'US',
+          value: '',
           priority: 10
+        },
+        fakeId2: {
+          id: 'fakeId2',
+          key: 'temp-hack',
+          value: '',
+          priority: 1
+        },
+        fakeId3: {
+          id: 'fakeId3',
+          key: 'temp-hack',
+          value: '',
+          priority: 1
         }
       }
     }
