+++ src/view/components/rule-engine/templates/rule-condition-component.tpl.html
@@ -9,8 +9,7 @@
       <div class="col-sm-11">
         <div class="row">
           <div class="col-sm-3">
-            <select class="form-control clause-selector" [value]="condition.conditionlet"
-                    (change)="setConditionlet($event.target.value)">
+            <select class="form-control clause-selector" [value]="conditionType.id" (change)="setConditionlet($event.target.value)">
               <option value="{{conditionType.id}}" *ng-for="var conditionType of conditionTypes; var i=index">
                 {{conditionType.name}}
               </option>
@@ -19,16 +18,19 @@
           <div class="col-sm-1">
             <h4 class="separator"></h4>
           </div>
-
           <cw-request-header-conditionlet
               *ng-if="conditionType?.id == 'UsersBrowserHeaderConditionlet'"
-              [header-key-value]="condition?.headerKeyValue"
               [comparator-value]="condition?.comparison"
               [comparison-values]="conditionValue"
-              (change)="updateAndStuff($event)"
+              (change)="conditionChanged($event)"
           >
           </cw-request-header-conditionlet>
           <cw-browser-conditionlet *ng-if="conditionType?.id == 'UsersBrowserConditionlet'"></cw-browser-conditionlet>
+          <cw-country-condition *ng-if="conditionType?.id == 'UsersCountryConditionlet'"
+                                [comparator-value]="condition?.comparison"
+                                [comparison-values]="conditionValue"
+                                (change)="conditionChanged($event)"
+          ></cw-country-condition>
 
           <div class="col-sm-1">
             <div class="operations">
